Added UI interaction with archive. Created VR mouse. Added mouse lock ability.
You can now lock the cursor to the center of the screen with an on
screen UI, simply press ESC to release the cursor.

Created a VR mouse that uses ray casting to interact with objects. Will
eliminate the issue of the normal mouse not functioning normally with
the Oculus Rift.

You can now hover over the the archive object now and it will get
illuminated using a sprite animation. This uses the new VR mouse to
work.